#5557 Fixes for Jackson 3 handling of null primitives

+++ stroom-util-shared/src/main/java/stroom/util/shared/ResultPage.java
@@ -110,7 +110,7 @@ public class ResultPage<T> implements Serializable {
                 }
 
                 final PageResponse pageResponse = new PageResponse(
-                        offset,
+                        (long) offset,
                         limited.size(),
                         (long) fullList.size(),
                         true);
@@ -259,8 +259,8 @@ public class ResultPage<T> implements Serializable {
                                 accumulator,
                                 new PageResponse(
                                         pageRequest != null
-                                                ? pageRequest.getOffset()
-                                                : 0,
+                                                ? (long) pageRequest.getOffset()
+                                                : 0L,
                                         accumulator.size(),
                                         counter,
                                         true));
